Namespace journey to EarlyYearsPayment::Provider
Change the journey from EarlyYearsPayment to EarlyYearsPayment::Provider since
there will be a second journey relating to the practitioner answers
