--- app/models/journeys/early_years_payment.rb
@@ -1,14 +1,0 @@
-module Journeys
-  module EarlyYearsPayment
-    extend Base
-    extend self
-
-    ROUTING_NAME = "early-years-payment"
-    VIEW_PATH = "early_years_payment"
-    I18N_NAMESPACE = "early_years_payment"
-    POLICIES = [Policies::EarlyYearsPayment]
-    FORMS = {
-      "claims" => {}
-    }
-  end
-end
