--- app/models/journeys/early_years_payment/eligibility_checker.rb
@@ -1,6 +1,0 @@
-module Journeys
-  module EarlyYearsPayment
-    class EligibilityChecker < Journeys::EligibilityChecker
-    end
-  end
-end
