+++ app/models/journeys/early_years_payment/provider/eligibility_checker.rb
@@ -1,0 +1,8 @@
+module Journeys
+  module EarlyYearsPayment
+    module Provider
+      class EligibilityChecker < Journeys::EligibilityChecker
+      end
+    end
+  end
+end
