+++ app/models/journeys/early_years_payment/provider/session.rb
@@ -1,0 +1,9 @@
+module Journeys
+  module EarlyYearsPayment
+    module Provider
+      class Session < Journeys::Session
+        attribute :answers, SessionAnswersType.new
+      end
+    end
+  end
+end
