+++ app/models/journeys/early_years_payment/provider/session_answers.rb
@@ -1,0 +1,11 @@
+module Journeys
+  module EarlyYearsPayment
+    module Provider
+      class SessionAnswers < Journeys::SessionAnswers
+        def policy
+          Policies::EarlyYearsPayment
+        end
+      end
+    end
+  end
+end
