+++ app/models/journeys/early_years_payment/provider/session_answers_type.rb
@@ -1,0 +1,7 @@
+module Journeys
+  module EarlyYearsPayment
+    module Provider
+      class SessionAnswersType < ::Journeys::SessionAnswersType; end
+    end
+  end
+end
