--- app/models/journeys/early_years_payment/session.rb
@@ -1,7 +1,0 @@
-module Journeys
-  module EarlyYearsPayment
-    class Session < Journeys::Session
-      attribute :answers, SessionAnswersType.new
-    end
-  end
-end
