--- app/models/journeys/early_years_payment/session_answers.rb
@@ -1,9 +1,0 @@
-module Journeys
-  module EarlyYearsPayment
-    class SessionAnswers < Journeys::SessionAnswers
-      def policy
-        Policies::EarlyYearsPayment
-      end
-    end
-  end
-end
