--- app/models/journeys/early_years_payment/session_answers_type.rb
@@ -1,5 +1,0 @@
-module Journeys
-  module EarlyYearsPayment
-    class SessionAnswersType < ::Journeys::SessionAnswersType; end
-  end
-end
