+++ app/views/early_years_payment/provider/landing_page.html.erb
@@ -1,6 +1,6 @@
 <div class="govuk-grid-row">
   <div class="govuk-grid-column-two-thirds">
-    <h1 class="govuk-heading-xl"><%= t("early_years_payment.journey_name") %></h1>
+    <h1 class="govuk-heading-xl"><%= t("early_years_payment_provider.landing_page.title") %></h1>
 
     <p class="govuk-body-l">
       Use this service to make an early years financial incentive payment claim
@@ -124,7 +124,7 @@
 
     <p class="govuk-body">
       You can cancel a claim or get help on this service by contacting us at
-      <%= govuk_link_to t("early_years_payment.feedback_email"), "mailto:#{t("early_years_payment.feedback_email")}", no_visited_state: true %>.
+      <%= govuk_link_to t("early_years_payment_provider.feedback_email"), "mailto:#{t("early_years_payment_provider.feedback_email")}", no_visited_state: true %>.
     </p>
 
     <p class="govuk-body">
