+++ spec/support/steps/journey_configuration.rb
@@ -2,6 +2,6 @@ def when_further_education_payments_journey_configuration_exists
   create(:journey_configuration, :further_education_payments)
 end
 
-def when_early_years_payment_journey_configuration_exists
-  create(:journey_configuration, :early_years_payment)
+def when_early_years_payment_provider_journey_configuration_exists
+  create(:journey_configuration, :early_years_payment_provider)
 end
